gadgetapp/offline_test/requirements.txt
psycopg2-binary==2.9.3
django==4.0.2
pytz==2022.1
